large-language-models/aws-faqs-chatbot/evaluation/requirements.txt
langchain==0.2.5
langchain-community==0.2.5
langchain-huggingface==0.0.3
chromadb== 0.5.3
beautifulsoup4==4.12.3
pandas==2.2.2
nltk==3.8.1
spacy==3.7.5
tqdm==4.66.4
numpy==1.26.4
scikit-learn==1.3.0
python-dotenv